<R11100> Major changes to setArticleState() to require a 'Published' state for any state beyond that, and to update the ebms_article_topic table.
Drastic simplification of getCurrentState(), now relying on the current='Y'
column and requiring that every state must have a topic_id, which was not
required by the old code but is now guaranteed.  The new code is vastly easier
to maintain.

Fixed an egregious bug in the class member version of setArticleState which no
one noticed before.

Reformatted comments for consistency.

Removed dead code.