Remove superfluous attribute missing its octothorpe
The citation queue form creation code was trying to tell Drupal
that a checkbox field should be unchecked by default. There is
an undocumented #checked property which if causes the checkbox
to be checked by default, but if the property is omitted the
checkbox is not checked when the form is rendered. So the property
in this case was unnecessary. However the property was named as
'checked' instead of '#checked' so Drupal expected the value indexed
by 'checked' to be an array, instead of an empty string. The latest
version of Drupal chokes when a string is passed to the form_builder
function as the &$element argument.

Fixes OCEEBMS-501.