Implement changed BMR report requirements (OCEEBMS-459)
In the original ticket, Robin asked us to rewrite all eight
of the "Citations by Status" reports. After I had done most
of the work for the request, Robin realized that the ticket
had been incorrectly written, and only one of the eight
reports was supposed to be modified (the "Board Member
Responses" report). So I very carefully backed out the changes
to the other seven reports (many other changes had been made
to the reports.inc file since the first commits for the ticket)
and asked the users to test all eight reports very carefully.

* Add a date range field set for the review submission date
* Remove the review cycle field
* Add a new Summary Topic column at the right of the report table
* Remove the Exclusion Reason, LOE, and Cycle columns